backend/requirements.txt
# Nexus Backend Service
# The missing piece that connects DefiLlama → Anomaly Detection → Oracle → Protection

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
web3==6.12.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
schedule==1.2.0
python-multipart==0.0.6
pydantic==2.5.0